refactor: 이벤트 참여 신청 API 분리
